final commit

+++ controllers/books/books_controller.go
@@ -2,21 +2,20 @@ package books
 
 import (
 	"github.com/bhanupratap1810/bookstore_users-api/controllers"
+	"github.com/bhanupratap1810/bookstore_users-api/controllers/middlewares"
 	"github.com/bhanupratap1810/bookstore_users-api/domain/books"
 	"github.com/bhanupratap1810/bookstore_users-api/utils/errors"
 	"github.com/gin-gonic/gin"
 	"net/http"
-	"strconv"
 )
 
-func getBookId(bookIdParam string) (int64, *errors.RestErr) {
-
-	bookId, bookErr := strconv.ParseInt(bookIdParam, 10, 64)
-	if bookErr != nil {
-		return 0, errors.NewBadRequestError("book id should be a number")
-	}
-	return bookId, nil
-}
+//func getBookId(bookIdParam string) (int64, *errors.RestErr) {
+//	bookId, bookErr := strconv.ParseInt(bookIdParam, 10, 64)
+//	if bookErr != nil {
+//		return 0, errors.NewBadRequestError("book id should be a number")
+//	}
+//	return bookId, nil
+//}
 
 //func getUserId(userIdParam string) (int64, *errors.RestErr) {
 //	userId, userErr := strconv.ParseInt(userIdParam, 10, 64)
@@ -27,7 +26,20 @@ func getBookId(bookIdParam string) (int64, *errors.RestErr) {
 //}
 
 func GetBooksHandler(service controllers.Service) gin.HandlerFunc {
-	return func(c * gin.Context) {
+	return func(c *gin.Context) {
+		userid, role, err1 := middlewares.GetUserIdAndRoleFromContext(c)
+		if err1 != nil {
+			restErr:=errors.NewUnauthorizedError("unauthenticated")
+			c.JSON(restErr.Status, restErr)
+			return
+		}
+
+		if userid==0 || role== "unemployed"{
+			restErr:=errors.NewBadRequestError("not permitted")
+			c.JSON(restErr.Status, restErr)
+			return
+		}
+
 		Books, err := service.BookServiceImpl.GetBooks()
 		if err != nil {
 			c.JSON(err.Status, err)
@@ -40,12 +52,28 @@ func GetBooksHandler(service controllers.Service) gin.HandlerFunc {
 func CreateBookHandler(service controllers.Service) gin.HandlerFunc {
 	return func(c *gin.Context) {
 		//service.UserServiceImpl.GetUser()
+
+		_, role, err1 := middlewares.GetUserIdAndRoleFromContext(c)
+		if err1 != nil {
+			restErr:=errors.NewUnauthorizedError("unauthenticated")
+			c.JSON(restErr.Status, restErr)
+			return
+		}
+
+
 		var book books.Book
 		if err := c.ShouldBindJSON(&book); err != nil {
 			restErr := errors.NewBadRequestError("invalid json body")
 			c.JSON(restErr.Status, restErr)
 			return
 		}
+
+		if role!="admin"{
+			restErr:=errors.NewForbiddenError("only admin can create books")
+			c.JSON(restErr.Status, restErr)
+			return
+		}
+
 		b, err := service.BookServiceImpl.CreateBook(book)
 		if err != nil {
 			c.JSON(err.Status, err)
@@ -57,7 +85,15 @@ func CreateBookHandler(service controllers.Service) gin.HandlerFunc {
 
 func UpdateBookHandler(service controllers.Service) gin.HandlerFunc {
 	return func(c *gin.Context) {
-		bookId, idErr := getBookId(c.Param("book_id"))
+
+		_, role, err1 := middlewares.GetUserIdAndRoleFromContext(c)
+		if err1 != nil {
+			restErr:=errors.NewUnauthorizedError("unauthenticated")
+			c.JSON(restErr.Status, restErr)
+			return
+		}
+
+		bookId, idErr := middlewares.GetId(c.Param("book_id"))
 		if idErr != nil {
 			c.JSON(idErr.Status, idErr)
 			return
@@ -65,6 +101,12 @@ func UpdateBookHandler(service controllers.Service) gin.HandlerFunc {
 		var book books.Book
 		if err := c.ShouldBindJSON(&book); err != nil {
 			restErr := errors.NewBadRequestError("invalid json body")
+			c.JSON(restErr.Status, restErr)
+			return
+		}
+
+		if role!="admin"{
+			restErr:=errors.NewForbiddenError("only admin can update books")
 			c.JSON(restErr.Status, restErr)
 			return
 		}
@@ -84,11 +126,26 @@ func UpdateBookHandler(service controllers.Service) gin.HandlerFunc {
 
 func DeleteBookHandler(service controllers.Service) gin.HandlerFunc {
 	return func(c *gin.Context) {
-		bookId, idErr := getBookId(c.Param("book_id"))
+
+		_, role, err1 := middlewares.GetUserIdAndRoleFromContext(c)
+		if err1 != nil {
+			restErr:=errors.NewUnauthorizedError("unauthenticated")
+			c.JSON(restErr.Status, restErr)
+			return
+		}
+
+		bookId, idErr := middlewares.GetId(c.Param("book_id"))
 		if idErr != nil {
 			c.JSON(idErr.Status, idErr)
 			return
 		}
+
+		if role!="admin"{
+			restErr:=errors.NewForbiddenError("only admin can delete books")
+			c.JSON(restErr.Status, restErr)
+			return
+		}
+
 		if err := service.BookServiceImpl.DeleteBook(bookId); err != nil {
 			c.JSON(err.Status, err)
 			return
@@ -99,7 +156,21 @@ func DeleteBookHandler(service controllers.Service) gin.HandlerFunc {
 
 func SearchBookByIdHandler(service controllers.Service) gin.HandlerFunc {
 	return func(c *gin.Context) {
-		bookId, idErr := getBookId(c.Param("book_id"))
+
+		userid, role, err1 := middlewares.GetUserIdAndRoleFromContext(c)
+		if err1 != nil {
+			restErr:=errors.NewUnauthorizedError("unauthenticated")
+			c.JSON(restErr.Status, restErr)
+			return
+		}
+
+		if userid==0 || role== "unemployed"{
+			restErr:=errors.NewBadRequestError("not permitted")
+			c.JSON(restErr.Status, restErr)
+			return
+		}
+
+		bookId, idErr := middlewares.GetId(c.Param("book_id"))
 		if idErr != nil {
 			c.JSON(idErr.Status, idErr)
 			return
