+++ controllers/users/users_controller.go
@@ -2,21 +2,20 @@ package users
 
 import (
 	"github.com/bhanupratap1810/bookstore_users-api/controllers"
+	"github.com/bhanupratap1810/bookstore_users-api/controllers/middlewares"
 	"github.com/bhanupratap1810/bookstore_users-api/domain/users"
 	"github.com/bhanupratap1810/bookstore_users-api/utils/errors"
 	"github.com/gin-gonic/gin"
 	"net/http"
-	"strconv"
 )
 
-func getUserId(userIdParam string) (int64, *errors.RestErr) {
-	userId, userErr := strconv.ParseInt(userIdParam, 10, 64)
-	if userErr != nil {
-		return 0, errors.NewBadRequestError("user id should be a number")
-	}
-	return userId, nil
-}
-
+//func getUserId(userIdParam string) (int64, *errors.RestErr) {
+//	userId, userErr := strconv.ParseInt(userIdParam, 10, 64)
+//	if userErr != nil {
+//		return 0, errors.NewBadRequestError("user id should be a number")
+//	}
+//	return userId, nil
+//}
 
 func CreateUserHandler(service controllers.Service) gin.HandlerFunc {
 	return func(c *gin.Context) {
@@ -37,15 +36,36 @@ func CreateUserHandler(service controllers.Service) gin.HandlerFunc {
 		//	c.JSON(saveErr.Status, saveErr)
 		//	return
 		//}
-		c.JSON(http.StatusCreated, u)
+		c.JSON(http.StatusCreated, u.Marshall(false))
 	}
 }
 
 func GetUserHandler(service controllers.Service) gin.HandlerFunc {
 	return func(c *gin.Context) {
-		userId, idErr := getUserId(c.Param("user_id"))
+
+		userid, role, err1 := middlewares.GetUserIdAndRoleFromContext(c)
+		if err1 != nil {
+			restErr:=errors.NewUnauthorizedError("unauthenticated")
+			c.JSON(restErr.Status, restErr)
+			return
+		}
+		//fmt.Println("========>", userid, role, err1 )
+
+		userId, idErr := middlewares.GetId(c.Param("user_id"))
 		if idErr != nil {
 			c.JSON(idErr.Status, idErr)
+			return
+		}
+
+		if userid==0 || role== "unemployed"{
+			restErr:=errors.NewBadRequestError("not permitted")
+			c.JSON(restErr.Status, restErr)
+			return
+		}
+
+		if role!="admin" && userid!=userId{
+			restErr:=errors.NewForbiddenError("only admin or the same user can get user details")
+			c.JSON(restErr.Status, restErr)
 			return
 		}
 
@@ -54,13 +74,28 @@ func GetUserHandler(service controllers.Service) gin.HandlerFunc {
 			c.JSON(getErr.Status, getErr)
 			return
 		}
-		c.JSON(http.StatusOK, user.Marshall(c.GetHeader("X-Public") == "true"))
+		isPublic:= userid!=userId
+		c.JSON(http.StatusOK, user.Marshall(isPublic))
 	}
 }
 
 func UpdateUserHandler(service controllers.Service) gin.HandlerFunc {
 	return func(c *gin.Context) {
-		userId, idErr := getUserId(c.Param("user_id"))
+		userid, role, err1 := middlewares.GetUserIdAndRoleFromContext(c)
+		if err1 != nil {
+			restErr:=errors.NewUnauthorizedError("unauthenticated")
+			c.JSON(restErr.Status, restErr)
+			return
+		}
+		//fmt.Println("========>", userid, role, err1)
+
+		if userid==0 || role== "unemployed"{
+			restErr:=errors.NewBadRequestError("not permitted")
+			c.JSON(restErr.Status, restErr)
+			return
+		}
+
+		userId, idErr := middlewares.GetId(c.Param("user_id"))
 		if idErr != nil {
 			c.JSON(idErr.Status, idErr)
 			return
@@ -68,6 +103,12 @@ func UpdateUserHandler(service controllers.Service) gin.HandlerFunc {
 		var user users.User
 		if err := c.ShouldBindJSON(&user); err != nil {
 			restErr := errors.NewBadRequestError("invalid json body")
+			c.JSON(restErr.Status, restErr)
+			return
+		}
+
+		if role!="admin" && userid!=userId{
+			restErr:=errors.NewForbiddenError("only admin or the same user can update the user details")
 			c.JSON(restErr.Status, restErr)
 			return
 		}
@@ -81,17 +122,38 @@ func UpdateUserHandler(service controllers.Service) gin.HandlerFunc {
 			c.JSON(err.Status, err)
 			return
 		}
-		c.JSON(http.StatusOK, result)
+		c.JSON(http.StatusOK, result.Marshall(false))
 	}
 }
 
 func DeleteUserHandler(service controllers.Service) gin.HandlerFunc {
 	return func(c *gin.Context) {
-		userId, idErr := getUserId(c.Param("user_id"))
+		userid, role, err1 := middlewares.GetUserIdAndRoleFromContext(c)
+		if err1 != nil {
+			restErr:=errors.NewUnauthorizedError("unauthenticated")
+			c.JSON(restErr.Status, restErr)
+			return
+		}
+
+		if userid==0 || role== "unemployed"{
+			restErr:=errors.NewBadRequestError("not permitted")
+			c.JSON(restErr.Status, restErr)
+			return
+		}
+
+		//fmt.Println("========>", userid, role, err1)
+		userId, idErr := middlewares.GetId(c.Param("user_id"))
 		if idErr != nil {
 			c.JSON(idErr.Status, idErr)
 			return
 		}
+
+		if role!="admin" && userid!=userId{
+			restErr:=errors.NewForbiddenError("only admin or the same user can delete the user details")
+			c.JSON(restErr.Status, restErr)
+			return
+		}
+
 		if err := service.UserServiceImpl.DeleteUser(userId); err != nil {
 			c.JSON(err.Status, err)
 			return
@@ -101,7 +163,21 @@ func DeleteUserHandler(service controllers.Service) gin.HandlerFunc {
 }
 
 func SearchUserHandler(service controllers.Service) gin.HandlerFunc {
-	return func(c * gin.Context) {
+	return func(c *gin.Context) {
+		_, roleByToken, err1 := middlewares.GetUserIdAndRoleFromContext(c)
+		if err1 != nil {
+			restErr:=errors.NewUnauthorizedError("unauthenticated")
+			c.JSON(restErr.Status, restErr)
+			return
+		}
+		//fmt.Println("========>", userid, roleByToken, err1)
+
+		if roleByToken!="admin" {
+			restErr:=errors.NewForbiddenError("only admin can search for the user details")
+			c.JSON(restErr.Status, restErr)
+			return
+		}
+
 		role := c.Query("Role")
 
 		Users, err := service.UserServiceImpl.SearchUser(role)
@@ -109,7 +185,8 @@ func SearchUserHandler(service controllers.Service) gin.HandlerFunc {
 			c.JSON(err.Status, err)
 			return
 		}
-		c.JSON(http.StatusOK, Users.Marshall(c.GetHeader("X-Public") == "true"))
+		//c.JSON(http.StatusOK, Users.Marshall(c.GetHeader("X-Public") == "true"))
+		c.JSON(http.StatusOK, Users.Marshall(true))
 	}
 }
